+++ src/types/Settings.ts
@@ -164,6 +164,10 @@ type Settings = {
     byBit: {
       timeout: number;
     };
+    binance: {
+      timeout: number;
+      maxBatchSize: number;
+    };
     cryptocompare: {
       apiKey: string;
       timeout: number;