+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - new `ByBitSpotFetcher`
+- new `BinanceFetcher`
 
 ## [7.28.6] - 2024-07-06
 ### Fixed