+++ CHANGELOG.md
@@ -6,13 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-### Fixed
-- fix conditions for feeds in multi fetchers
-- fix conversion to number in Uniswap V3 fetcher and fix support for unsuccessful prices
 
+## [7.29.0] - 2024-07-12
 ### Added
 - new `ByBitSpotFetcher`
 - new `BinanceFetcher`
+
+### Fixed
+- fix conditions for feeds in multi fetchers
+- fix conversion to number in Uniswap V3 fetcher and fix support for unsuccessful prices
 
 ## [7.28.6] - 2024-07-06
 ### Fixed